Initial Version
The Twitch Stream Status plugin takes a list of casters (defined in the dashboard) and displays those live as well as their currently listed game/activity. Both the caster's name and the game are clickable. Clicking the caster's name will take you to their channel (duh?) and clicking the game will take you to the Twitch.tv directory showing everyone streaming that game.

There is no caching, and the admin page needs to be a bit AJAX'ified so that you can add more than one streamer at a time.